+++ lib/inspector_controller.dart
@@ -1,21 +1,33 @@
 import 'package:flutter/material.dart';
+import 'package:requests_inspector/show_inspector_on_enum.dart';
 import 'package:shake/shake.dart';
 
 import 'request_details.dart';
 
 ///Singleton
 class InspectorController extends ChangeNotifier {
-  factory InspectorController([bool enabled = false]) =>
-      _singleton ??= InspectorController._internal(enabled);
+  factory InspectorController({
+    bool enabled = false,
+    ShowInspectorOn showInspectorOn = ShowInspectorOn.Shaking,
+  }) =>
+      _singleton ??= InspectorController._internal(
+        enabled,
+        showInspectorOn,
+      );
 
-  InspectorController._internal(bool enabled) : _enabled = enabled {
-    if (_enabled)
-      _shakeDetector = ShakeDetector.autoStart(onPhoneShake: _showInspector);
+  InspectorController._internal(
+    bool enabled,
+    ShowInspectorOn showInspectorOn,
+  )   : _enabled = enabled,
+        _showInspectorOn = showInspectorOn {
+    if (_enabled && _allowShaking)
+      _shakeDetector = ShakeDetector.autoStart(onPhoneShake: showInspector);
   }
 
   static InspectorController? _singleton;
 
   late final bool _enabled;
+  late final ShowInspectorOn _showInspectorOn;
   late final ShakeDetector _shakeDetector;
 
   final pageController = PageController(
@@ -33,6 +45,10 @@ class InspectorController extends ChangeNotifier {
   int get selectedTab => _selectedTab;
   List<RequestDetails> get requestsList => _requestsList;
   RequestDetails? get selectedRequested => _selectedRequested;
+  bool get _allowShaking => [
+        ShowInspectorOn.Shaking,
+        ShowInspectorOn.Both,
+      ].contains(_showInspectorOn);
 
   set selectedTab(int value) {
     if (_selectedTab == value) return;
@@ -47,7 +63,7 @@ class InspectorController extends ChangeNotifier {
     notifyListeners();
   }
 
-  void _showInspector() => pageController.jumpToPage(1);
+  void showInspector() => pageController.jumpToPage(1);
 
   void hideInspector() => pageController.jumpToPage(0);
 
@@ -59,7 +75,7 @@ class InspectorController extends ChangeNotifier {
 
   @override
   void dispose() {
-    _shakeDetector.stopListening();
+    if (_allowShaking) _shakeDetector.stopListening();
     super.dispose();
   }
 }