+++ lib/show_inspector_on_enum.dart
@@ -1,0 +1,5 @@
+enum ShowInspectorOn {
+  LongPress,
+  Shaking,
+  Both,
+}